+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+## [0.0.2] - 2024-10-10
+
+- Added named hybrid and hybrid formula handling
+- Added original authorship and combination authorship handling
 - Added et_al_cutoff parameter to authorship formatting
 
 ## [0.0.1] - 2024-03-27